'v3.7.2.1'

+++ 100_core/src/gplx/core/envs/Env_.java
@@ -52,7 +52,7 @@ public class Env_ {
 	public static int TickCount_elapsed_in_frac(long time_bgn) {return (int)(Env_.TickCount() - time_bgn);}
 	public static long TickCount_Test = -1;	// in milliseconds
 	public static void TickCount_normal() {TickCount_Test = -1;}
-	public static long System_cpu_count() {return Runtime.getRuntime().availableProcessors();}	
+	public static int System_cpu_count() {return Runtime.getRuntime().availableProcessors();}	
 	public static long System_memory_max() {return Runtime.getRuntime().maxMemory();}	
 	public static long System_memory_total() {return Runtime.getRuntime().totalMemory();}	
 	public static long System_memory_free() {return Runtime.getRuntime().freeMemory();}
